refactor(mail): group services into a services folder
